:dizzy: Adding the LogInScreen to the connections stack
