+++ screens/connections/SignUpScreen.js
@@ -1,14 +1,23 @@
-import { TextButton } from "../../components/buttons/TextButton";
-import { View, Text, Separator } from "../../constants/Themed";
-import { TextInputPassword, TextInputGlobal } from "../../components/TextInput";
-import { global, textInput, color } from "../../style/styles";
-import Colors from "../../constants/Colors";
 import { useState } from "react";
-import { signUp, signUpCode, signIn } from "../../API/Login";
 import { SafeAreaProvider } from "react-native-safe-area-context";
 
-export function SignUpScreen({ navigation }) {
+import { signUp, signUpCode, signIn } from "../../API/Login";
 
+import { View, Text } from "../../constants/Themed";
+import { global, textInput } from "../../style/styles";
+import Colors from "../../constants/Colors";
+
+import { TextButton } from "../../components/buttons/TextButton";
+import { HideTextInput, TextInput } from "../../components/TextInput";
+
+/**
+ * @class display Login screen
+ * @export
+ * 
+ * @description A function that returns a View to sign in or sign up.
+ * @return {HTML} 
+ */
+export function SignUpScreen({ navigation }) {
     const [email, setEmail] = useState('')
     const [password, setPassword] = useState('')
     const [code, setCode] = useState("")
@@ -42,8 +51,8 @@ export function SignUpScreen({ navigation }) {
     return (
         <View style={global.container}>
             <View style={global.middleContainer}>
-                <TextInputGlobal autoCapitalize='none' style={[textInput.global, { borderColor: Colors(checkMailError() ? 'ErrorRed' : 'dedalBlue') }]} placeholder='Email' onChangeText={setEmail} value={email}></TextInputGlobal>
-                <TextInputPassword autoCapitalize='none' style={[textInput.global, { borderColor: Colors(checkPasswordError() ? 'ErrorRed' : 'dedalBlue') }]} placeholder="password" onChangeText={setPassword} value={password}></TextInputPassword>
+                <TextInput autoCapitalize='none' style={[textInput.global, { borderColor: Colors(checkMailError() ? 'ErrorRed' : 'dedalBlue') }]} placeholder='Email' onChangeText={setEmail} value={email} />
+                <HideTextInput autoCapitalize='none' style={[textInput.global, { borderColor: Colors(checkPasswordError() ? 'ErrorRed' : 'dedalBlue') }]} placeholder="password" onChangeText={setPassword} value={password} />
                 {!state ?
                     <View style={{ paddingTop: 20 }}>
                         <Text>Votre mot de passe doit contenir au moins : </Text>
@@ -52,7 +61,7 @@ export function SignUpScreen({ navigation }) {
                     </View>
                     :
                     <>
-                        <TextInputGlobal autoCapitalize='none' style={textInput.global} placeholder="Code" onChangeText={setCode} value={code}></TextInputGlobal>
+                        <TextInput autoCapitalize='none' style={textInput.global} placeholder="Code" onChangeText={setCode} value={code} />
                         <View style={{ paddingTop: 20 }}>
                             <Text>{error ? "une erreur est subvenu veillez réésailler" : "Un code de verification vous à était envoyé sur voter adresse mail."}</Text>
                         </View>
